Extract logic to create a Changeset object to ChangesetFactory
