+++ scribe-infrastructure/src/main/java/eu/xenit/contentcloud/scribe/infrastructure/changeset/dto/ProjectDto.java
@@ -1,8 +1,10 @@
 package eu.xenit.contentcloud.scribe.infrastructure.changeset.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
+@AllArgsConstructor
 public class ProjectDto {
     private String name;
     private String organization;
